services/ai-service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
gunicorn>=21.2
requests>=2.31
python-dotenv>=1.0
